workflow: let mariner process multiple times
When do_mariner() is called, the IMAGES/VOLUMES/NETWORKS/COMMANDS define
the schema to be processed, resulting in an output Makefile that is
sourced back in, and additional attributes defined by the processing.
There are cases where we want to do this before defining other rules, so
that we can leverage the processing of earlier elements. But we could not
add new schema and not generate new Makefile output, because the
do_mariner() ship had already sailed at that point.

With this change, do_mariner() can be run multiple times, and will
generate (and source back in) multiple Makefile outputs, so the schema
can be digested in incremental chunks.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>