workflow: add hostname lookup to git API
The git repo implements an efficient ekpubhash-to-hostname lookup, but
here we add a much-less-efficient reverse lookup on hostname suffix, and
corresponding API.

This also takes the directory-ply definitions out of common.sh and puts
them into a common_defs.sh that is actually committed to the git repo. In
this way, clones of the repo will have those definitions too.

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>