workflow: add ephemeral/local rule generation, use by default
For constructs that require multiple steps, where in each intermediate
state the workflow would generate different target dependencies than it
generates in the initial state, a higher-level workflow should be
defined that executes "make" calls for low-level workflow constructs. The
states and rules for this should be explicitly decoupled from the lower
level workflow. An example of this requirement is the automation of test
cases, where the lower-level dependency-driven rules are used to bring up,
exercise, and then tear down images, volumes, etc from that lower-level
workflow.

In a typical workflow, Mariner-generated rules for images, verbs, volumes,
and networks will be based on their current state. E.g. "make setup-foo"
would move "foo" to the "setup" state. (The state name is often the past
tense of the verb it executes. Or in the case of async verbs, there are two
past tense states per verb, one for 'started' and one for 'done'. Services
are implemented this way, as async verbs named "run".) Such state changes
often cause state changes in related objects, including creating or
destroying them, or starting (and/or ending) the execution of verbs, etc.
E.g. the rules for a "stop" or "reset" action/state might be influenced by
what has or hasn't already been created, what has or hasn't already run (or
is still running), and so on. An automation workflow that manipulates these
stateful elements in a to-and-fro that can't (easily) be described by a
static/unchanging dependency DAG would instead need to be described as a
distinct, higher-level workflow, which could then invoke multiple "make"
steps of the lower-level workflow in the recipes of that higher-level
workflow.

For that to work, this patch is needed! It allows the entire makefile
generation process to be made local and/or ephemeral, rather than global
(as it previously was; generating into workflow/crud/Makefile.out.*). The
default is now local and ephemeral: each make invocation generates its
rules to a "mktemp -d"'d directory, and deletes them once rule-generation
is over. MARINER_MKOUT_STICKY can be set to suppress deletion of these
files, for debugging. In this way, multiple invocations can coexist that
generate different dependencies and rules without clobbering each other,
whether they are at different layers or running in parallel (or both).

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>