attest-enroll: improve error-handling
When an individual genprog breaks (because of code or configuration), it
is easier to find if the enrollment transaction bails out right where the
error occurred. Instead, having enrollments continue and succeed (but be
incomplete) causes future assumptions to fail, when the root cause may be
far less obvious.

This patch changes attest-enroll to return success only if _all_ of the
implied actions succeed.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>