attest-enroll: allow genprogs to generate more
Previously, a genprog was expected to produce a single line of output (to
stdout) specifying "<public|sensitive> [files...]" or "skip". This allows
the genprog to produce multiple files, but only if they are of the same
type (public or sensitive). With this change, attest-enroll will process
multiple lines of output, one at a time.

Signed-off-by: Geoffrey Thorpe <geoffrey@twosigma.com>