tpm2-attest: distinguish "unrecognized TPM" from other errors
This now captures the status code as well as the output when attesting. If
the status code is a 404, tpm2-attest exits with 2 rather than 1, allowing
a caller/user to distinguish. (This is necessary when implementing retry
logic that should wait for enrollments to propagate but not for other
error conditions to resolve themselves with time.)

Signed-off-by: Geoff Thorpe <geoffrey@twosigma.com>